VIAM-115 hmac-mapperextension: adding additional pseudonyms used for other clients
adds another OIDC protocollmapper which calculates the pseudonyms of a given list of clients.
It uses the original client mapper configuration for the given clients.
Checks Clients on Save.
If Client is removed later client is skipped only with a warning in the log.

Todo Unit and postman tests.

FWUV-126